requirements.txt
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-oauth==1.0.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
oauthlib==3.2.2
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.3.0
pyparsing==3.2.0
PyYAML==6.0.2
ratelimit==2.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
soupsieve==2.6
uritemplate==4.1.1
urllib3==2.3.0
